And the LORD said, "Thou shalt not break WRFDA"
And it was so.

WTF now has WRFDA test capabilities. Just a very simple 3DVAR test for now (though there is opportunity for expansion), and since diffwrf is not built for WRFDA the bit-for-bit test will always report success regardless of actual success (though there is opportunity to fix this in the future).



git-svn-id: https://svn-wrf-wtf.cgd.ucar.edu/trunk@212 23d9a9be-b5ef-4b6d-934b-143d34859f8a